Create new fragment class for the standard videos and separate the logic from CourseUnitVideoFragment.
